WIP – refactor code
