+++ README.md
@@ -3,13 +3,3 @@
 # cra-with-api
 
 Boilerplate for Create-React-App with an Express.js API
-
-# Installation
-
-```
-$ npm i
-$ mv .env.example .env # You need an API_KEY from our company passwort manager
-$ npm run dev # for development
-# OR
-$ npm run build && npm start # for production
-```
