--- client/src/services/getProfile.js
@@ -1,3 +1,0 @@
-export default function getProfile() {
-  return fetch('/me').then((res) => res.json())
-}
